assignment-7-serverless/function-http-trigger/requirements.txt
google-cloud-storage
functions-framework==3.*